fix: Retry configuring mvbox on IMAP reconnect if mvbox_move is enabled explicitly
If `mvbox_move` isn't enabled explictly, but only by default, better not to retry configuring it to
avoid excessive work, but if it's enabled explicitly by the user and actually can't be configured,
the user should disable it. But anyway not being able to configure the mvbox doesn't prevent Delta
Chat from working just leaving all mail in Inbox.