fix: Order of messages if Sentbox is synced before Inbox
This fixes a Gmail-like scenario when outgoing messages are saved to Sentbox as well and it is
fetched before Inbox by chance, by adding a new `OutRcvd` message state for received outgoing
messages so that they can mingle with fresh incoming ones. As for `OutPending`, `OutDelivered`
etc. messages, they are sent locally by the user and there's no need to make them more noticeable
even if they are newer.

All APIs still return `OutDelivered` instead of `OutRcvd` for compatibility.