fix: Emit MsgsNoticed on receipt of an IMAP-seen message
`imap::Session::sync_seen_flags()` emits `MsgsNoticed` for existing messages seen on other devices,
so `receive_imf` should do the same when it receives a seen message. Otherwise a multi-device user
may see a new message notification on device A, just swipe it, then see another new message
notification and mark it as read, and when a device B goes online, it will show a notification for
the first message, and it won't be removed because `MsgsNoticed` isn't emitted. I have checked this
with my DC Android and Desktop. With this fix the notification can be removed if UIs handle
`MsgsNoticed` accordingly, though currently i can't achieve this behaviour in Desktop.

Also add blocked chat's messages and silent group changes messages as `InNoticed` instead of
`InSeen`, they don't need to cause `MsgsNoticed` events. This doesn't change any UX, checked this in
Desktop.