fix: try_many_times: retry 5 times instead of 60
If 5 times is not enough, 60 will probably not be enough either.

This is mainly an attempt of improving the situation with
https://github.com/deltachat/deltachat-desktop/issues/3959.

The `remove_account` RPC call would make the RPC server
stop responding to all other requests,
which is basically equivalent to a one minute hang.

Additionally, `try_many_times` appears to be unnecessary after
https://github.com/chatmail/core/pull/5814#issuecomment-2257164502.