fix: Keep email email contact in 1:1 chat if PGP contact has a different addr
If a contact did AEAP in the past, then it can happen that there are multiple contacts with the same email address but different keys.
In this case, we should make sure that we don't send messages to a different email address all of a sudden,
and that there aren't two 1:1 chats with the same contact.

Therefore, if the PGP contact has a different email address than the
email contact, we keep the email contact in 1:1 chats.