feat: If any group member is in RESET peerstate, downgrade to Ad-Hoc group (#6905)
I agree with holger that it's most important to allow communication in
mixed/ambiguous groups, even if you can't modify the member list
anymore.

So, if any member of a group is in RESET peerstate, this downgrades the
whole group to an Ad-Hoc group, rather than removing this member.
Except for chatmail accounts,
where we ignore the Reset state already before PGP-contacts,
so, it's ignored in the migration, too.

Broadcast lists keep the current behavior.

Fix #6908