doc: Adapt the comment on `MessageObject.showPadlock` (#6925)
We want to remove the padlock-icon from messages, and instead show an
icon if the message was unencrypted:
https://github.com/deltachat/deltachat-ios/pull/2746/

Now, it's confusing for implementers that there is a field `showPadlock`
in the jsonrpc. This PR adapts the comment.

Two alternatives to this PR would be:
- Deprecate `showPadlock`, and introduce a field `wasEncrypted`
- Deprecate `showPadlock`, and introduce a field `showLetterIcon`,
which is the inverse of `showPadlock`.