fix: Also migrate messages in PGP-contacts migration (#6921)
Previously, messages were not rewritten. This meant that all messages
stayed with the old email-identified contact.
#6916 made it very obvious that all messages sent into a group before
the PGP-contacts migration got the email avatar.

With this PR, all encrypted messages are rewritten to the PGP-contact
identified by the current autocrypt key. It is not possible to find out
which key was actually used to sign the message.

---------

Co-authored-by: iequidoo <117991069+iequidoo@users.noreply.github.com>