fix: Add encrypted ad-hoc groups
Allow encrypted ad-hoc groups in addition to unencrypted:
- Create an encrypted ad-hoc group from an encrypted multi-participant non-chat message. It's not
  completely secure because we don't know the exact keys the message is encrypted to and we use an
  heuristic based on `contacts.last_ssen`, but it's more secure than creating an unencrypted group
  and adds an interoperable group encryption in some way.
- Assign encrypted non-chat messages to groups with matching name and members. This is useful when
  e.g. the user was removed from a group and the re-added by a non-chat message; in this case we
  can't rely on assigning by References. Another case is reordered messages. This is safe because
  such messages don't modify the group members, so the next outgoing message is encrypted to the
  same keys.

This doesn't change the existing ad-hoc groups db migration because probably it's too late. No tests
are broken, `create_group_ex()` preserves behavior and encryption status never changes once the
group is created, so UIs should continue to work.

Also this fixes group members for ad-hoc groups created by encrypted messages: currently, if
somebody replies encrypted, their contact is duplicated with a key-contact and the number of members
is reported wrongly. NB: This also can be fixed by assigning the message to the key-contact, but
adding the address-contact to the chat instead, so the message will be from a non-member
technically, but "Reply Privately" will encrypt as currently.

Known bugs:
- UI's "QR Invite Code" button doesn't work, an error is shown. It can't work w/o a Group Id.
- When adding/removing members, non-chat contacts receive weird system messages. OTOH currently
  that's not possible at all, so it's not a regression and the problem exists anyway for Delta Chat
  groups with non-chat contacts.