fix: Make "end-to-end encryption not setup" message work again
Reported by adbenitez in the release coordination group:

The "end-to-end encryption not setup" message didn't appear anymore when
trying to send an unencrypted message with a chatmail account. This was
because on the server side, the casing of the error message changed
("Inalid unencrypted mail" -> "Invalid Unencrypted mail").

This PR ignores the casing.