feat: do not replace avatar for address-contact profiles
Key-contacts and address-contacts do not appear
in the same lists in the UI, so there is no need
to distinguish between the two.

Letter icon is renamed from "address contact icon"
to "unencrypted chat icon" to make it clear
that the icon distinguishes unencrypted chats
from encrypted chats as it is also used for
unencrypted group chats and is most helpful
in the chatlists where a mix of encrypted
and unencrypted chats is displayed.