fix: check for Chat-Version header before marking the contact as a bot
Contacts are marked as a bot if incoming message from a contact
has `Auto-Submitted: auto-generated` header.
However, such message can be generated by the server,
e.g. as an auto-reply.

To prevent accidentally marking the contact as a bot
when auto-generated message is sent by the server,
we need to check for `Chat-Version` header
to make sure that the message is generated by the client.

This is already how it is documented in `MimeMessage.is_bot`
field documentation comment, but was not checked in the implementation.