api!: Contact::get_color(): Take color for SELF from Config::Selfcolor
4010c60e7b6d00846fe23d688a4933ac4e9ca2df "feat: use key fingerprints for color generation" changes
colors for contacts including SELF. Even before i was asked whether there's a way to change the
self-color. The user can set an avatar, but anyway the self-color is visible in e.g. replies to
outgoing messages. So we need a way to change/revert the self-color. This color doesn't need to be
limited to some "contrast color set" though. So, introduce `Config::Selfcolor` for that, we can add
a UI later. Users can set it meanwhile using `deltachat-repl`, it's not supposed to be a frequent
operation anyway.