feat: Be more generous with marking contacts as verified for now
Context: PR #7116 is backwards-incompatible with versions older than v2.20, and since the release hasn't reached all users yet, we currently can't release from main; for
details see #7326.

Issue #7326 explains how we can make this less breaking, but this only
works if many contacts are verified. So, this PR here proposes to
postpone the stricter rules for who is verified a bit:

- Set verification timeout for invite codes to 1 week (this is still
  stricter than no timeout at all, which we had in the past)
- Don't reset indirect verifications yet (we should reset them in
  a few months then, but this is just so that we can release from main again,
  and release e.g. channels).