fix: Use Chat-List-Id rather than List-Id header
Without this commit, there is this problem:

In a multi-device setup where Alice1 uses DC v2.22.0 and Alice2 uses this PR, and Alice2 performs a securejoin to add someone to a channel, Alice1 will be stuck in an infinite loop similar to https://github.com/chatmail/core/issues/7290 because `observe_securejoin_on_other_device()` fails.

The solution is to use a header different than List-Id, so that old versions of DC don't recognize it as a mailing list.